requirements.txt
mypy==0.812
mypy-extensions==0.4.3
Pillow==8.2.0
Send2Trash==1.5.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
